requirements.txt
google-api-core==2.15.0
google-api-python-client==2.116.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.13.0
google-cloud-core==2.4.1
google-cloud-dataform==0.5.4
google-cloud-pubsub==2.18.4
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
google-cloud-secret-manager==2.18.2
grpc-google-iam-v1==0.13.0
numpy==1.26.2
pandas==2.1.4
requests==2.31.0
tqdm==4.66.1
beautifulsoup4
streamlit==1.40.2
stqdm
db-dtypes